feat: better split tests
